Add multiple aircraft capability to OpenAirClim
This commit introduces the capability to consider the climate
impact of multiple different aircraft designs within the same
emission inventory.

The aircraft identifiers are defined in the
config file under `[aircraft]` and must correspond to the values
in the emission inventories (data variable "ac"). If the data
variable "ac" is not present in the emission inventories, the
identifier "DEFAULT" is used, which must be defined in the
config file. The results are provided along the new coordinate
"ac" and include the identifier "TOTAL".

The current main limitation is that due to non-linear effects,
contrails cannot currently be calculated. The attribution
methdology is work in progress.

See: #16 and #69