Add debugging details to oac output
For debugging and diagnosing results, it is useful
to have more information about the OpenAirClim
version used, the log and the config file used. This
commit adds the OpenAirClim version number to
the output file, copies the config file to the results
folder and moves the debug file there too. In
future, debugging and logging will be properly
centralised - see #63.

Closes #66