Add contrail capability for multiple aircraft
This commit re-introduces contrail capability into the multiple aircraft
framework. Contrails can only be simulated if a single aircraft is
present in the emission inventories since the attribution methodology
is still work in progress. If multiple aircraft identifiers are present
in the emission inventories, OpenAirClim will throw a ValueError. This
is temporary and will be removed once the attribution methodology
has been developed.

The "TOTAL" results of the multiple aircraft framework match those of
OpenAirClim v0.10.0. Minor differences are present in RF_CH4 (and by
extension RF_PMO), most likely due to the accuracy of the interpolation.

Closes #16
See also: #69 #70