Fix error for empty aircraft identifiers
During testing, it was found that if an aircraft was defined in the
config file, but was not present in the emission inventories,
OpenAirClim would fail. This commit re-defines modifies the
calculation of `full_inv_dict` and re-defines `ac_lst` to only consider
aircraft identifiers that are present in the emission inventories.

I also fixed the longitude check for the emission inventories, "lon" was
assigned as a coordinate rather than a data variable in `read_netcdf.py`