Add environment files and installation documentation for calc_h2.py
The calc_h2.py module depends on the python package tensorflow. For Windows systems, special requirements have to be met for the installation of tensorflow. The environment files create dedicated conda environments for the execution of calc_h2.py commands. Installation instructions are added to the documentation in hydrogen.md.