Add capability for 3D-resolved ISS
Originally, ISS was overlapped between 200 and 250 hPa. From the new
CCMod simulations, we now have the ice supersaturation fraction for all
pressure levels. Therefore, this commit removes the vertical sum in the
CFDD calculation and adds a new resp_cont with the 3D-resolved ISS. The
cccov is then randomly overlapped. This may be amended again when
the attribution functionality is introduced.

Resolves: #68